requirements.txt
Django==4.2.21
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7